+++ src/yb/client/txn-test-base.cc
@@ -270,7 +270,7 @@ void TransactionTestBase<MiniClusterType>::VerifyData(
 
 template <>
 bool TransactionTestBase<MiniCluster>::HasTransactions() {
-  for (int i = 0; i != cluster_->num_tablet_servers(); ++i) {
+  for (size_t i = 0; i != cluster_->num_tablet_servers(); ++i) {
     auto* tablet_manager = cluster_->mini_tablet_server(i)->server()->tablet_manager();
     auto peers = tablet_manager->GetTabletPeers();
     for (const auto& peer : peers) {
@@ -302,7 +302,7 @@ template <>
 bool TransactionTestBase<MiniCluster>::CheckAllTabletsRunning() {
   bool result = true;
   size_t count = 0;
-  for (int i = 0; i != cluster_->num_tablet_servers(); ++i) {
+  for (size_t i = 0; i != cluster_->num_tablet_servers(); ++i) {
     auto peers = cluster_->mini_tablet_server(i)->server()->tablet_manager()->GetTabletPeers();
     if (i == 0) {
       count = peers.size();
